feat: implementa funcionalidade de carregamento de livros por gêneros com integração e melhorias na lista de desejos
