+++ public/js/pages/genero.js
@@ -1,4 +1,7 @@
 import { getGeneros } from "../api/genero.js";
+import { searchBooks } from "../api/livro.js";
+import { renderBooks, renderSkeletons } from "../utils/renderBooks.js";
+import { carregarListaDesejos, configurarBotoesFavoritos } from "../utils/wishlist-utils.js";
 
 export async function carregarGeneros(selectClass) {
   try {
@@ -8,8 +11,8 @@ export async function carregarGeneros(selectClass) {
       return;
     }
 
-    const response = await getGeneros(); 
-    
+    const response = await getGeneros();
+
     if (response.status === "success") {
       console.log("Gêneros carregados com sucesso:", response);
       response.data.forEach((genero) => {
@@ -28,4 +31,79 @@ export async function carregarGeneros(selectClass) {
 
 document.addEventListener("DOMContentLoaded", () => {
   carregarGeneros("genero"); // Classe do <select> de gêneros
+});
+
+async function carregarLivrosPorGenero() {
+  const container = document.getElementById("generos-container");
+
+  try {
+    const generosResponse = await getGeneros();
+    if (generosResponse.status !== "success") {
+      throw new Error("Erro ao carregar gêneros.");
+    }
+
+    const generos = generosResponse.data;
+
+    // Carregar lista de desejos
+    const favoritos = await carregarListaDesejos();
+
+    for (const genero of generos) {
+      // Cria uma seção para o gênero
+      const generoSection = document.createElement("section");
+      generoSection.classList.add("genero-section");
+
+      // Adiciona o título do gênero
+      const generoTitulo = document.createElement("h2");
+      generoTitulo.textContent = genero.nome;
+      generoTitulo.classList.add("heading-secondary");
+
+      // Cria o contêiner para os livros
+      const livrosGrid = document.createElement("div");
+      livrosGrid.classList.add("grid", "grid--4-cols");
+
+      // Renderiza os esqueletos enquanto os livros são carregados
+      renderSkeletons(livrosGrid, 4);
+
+      // Adiciona o título e os esqueletos à seção
+      generoSection.appendChild(generoTitulo);
+      generoSection.appendChild(livrosGrid);
+
+      // Adiciona a seção ao contêiner principal
+      container.appendChild(generoSection);
+
+      // Busca os livros do gênero
+      const livrosResponse = await searchBooks("", genero.id);
+      if (livrosResponse.status === "success") {
+        const livros = livrosResponse.data;
+
+        if (livros.length > 0) {
+          // Substitui os esqueletos pelos livros reais
+          livrosGrid.innerHTML = ""; // Limpa os esqueletos
+          renderBooks(livrosGrid, livros);
+
+          // Configurar botões de favoritos
+          configurarBotoesFavoritos(favoritos, ".btn-favorito");
+        } else {
+          // Remove a seção se não houver livros
+          container.removeChild(generoSection);
+        }
+      } else {
+        // Remove a seção se houver erro ao carregar os livros
+        container.removeChild(generoSection);
+        console.warn(`Erro ao carregar livros para o gênero ${genero.nome}: ${livrosResponse.message}`);
+      }
+    }
+  } catch (error) {
+    console.error("Erro ao carregar livros por gênero:", error);
+    container.innerHTML = "<p>Erro ao carregar os livros. Tente novamente mais tarde.</p>";
+  }
+}
+
+document.addEventListener("DOMContentLoaded", () => {
+  const container = document.getElementById("generos-container");
+  if (container) {
+    carregarLivrosPorGenero();
+  } else {
+    console.warn("Elemento 'generos-container' não encontrado no DOM.");
+  }
 });
