+++ view/admin.html
@@ -24,7 +24,7 @@
       <nav class="main-nav">
         <ul class="main-nav-list">
           <li><a href="home.html" class="main-nav-link navegacao">Home</a></li>
-          <li><a href="#" class="main-nav-link navegacao">Gêneros</a></li>
+          <li><a href="generos.html" class="main-nav-link navegacao">Gêneros</a></li>
           <li><a href="#" class="main-nav-link navegacao">Catálogo</a></li>
 
           <div class="dropdown">
