+++ view/home.html
@@ -35,7 +35,7 @@
 
           <li><input type="text" placeholder="Pesquisar Livro..." /></li>
           <li><button class="btn">Buscar</button></li>
-          <li><a href="#" class="main-nav-link navegacao">Gêneros</a></li>
+          <li><a href="generos.html" class="main-nav-link navegacao">Gêneros</a></li>
           <li><a href="#" class="main-nav-link navegacao">Catálogo</a></li>
 
           <li>
